# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Contoh API documentation",
        "description": "Contoh API doc menggunakan OpenAPI/Swagger",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "choirudin.emchagmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "paths": {
        "/api/greet": {
            "get": {
                "tags": [
                    "greeting"
                ],
                "summary": "Returns a Sample API response",
                "description": "A sample greeting to test out the API",
                "operationId": "greet",
                "parameters": [
                    {
                        "name": "firstname",
                        "in": "query",
                        "description": "nama depan",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lastname",
                        "in": "query",
                        "description": "nama belakang",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/gallery": {
            "get": {
                "tags": [
                    "gallery"
                ],
                "summary": "Returns a list of gallery images",
                "description": "Returns a paginated list of gallery images where pictures are not empty",
                "operationId": "getGallery",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "picture"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Post title",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Post description",
                                        "type": "string"
                                    },
                                    "picture": {
                                        "description": "Post image",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/postPhoto": {
            "post": {
                "tags": [
                    "Post Photo"
                ],
                "summary": "Upload a photo",
                "description": "Endpoint to upload a photo with title and description",
                "operationId": "postPhoto",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Title of the photo"
                                    },
                                    "description": {
                                        "description": "Description of the photo"
                                    },
                                    "image": {
                                        "description": "Image file to be uploaded",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "Successful operation"
                    }
                }
            }
        }
    }
}